+++ server/src/test/scala/io/delta/standalone/internal/JsonPredicateFilterUtilsSuite.scala
@@ -19,8 +19,6 @@ package io.delta.standalone.internal
 import io.delta.standalone.internal.actions.AddFile
 import org.scalatest.FunSuite
 
-import io.delta.sharing.server.util.JsonUtils
-
 class JsonPredicateFilterUtilsSuite extends FunSuite {
 
   import JsonPredicateFilterUtils._
